fix: adjust eps value in RMSNorm and improve attention stability
- Change RMSNorm eps from 1e-6 to 1e-5 for better numerical stability
- Cast attention scores to float before softmax to prevent overflow
- Add causal mask to model inference and fix code formatting

+++ transformer.py
@@ -104,8 +104,8 @@ class Attention(nn.Module):
 
         # 应用softmax和dropout
         attn_weights = F.softmax(
-            scores, dim=-1
-        )  # (batch_size, heads, seq_len, seq_len)
+            scores.float(), dim=-1
+        ).to(scores.dtype)  # (batch_size, heads, seq_len, seq_len)
         attn_weights = self.dropout(attn_weights)
 
         # 加权求和
